new support

+++ src/Generator.php
@@ -9,15 +9,18 @@ class Generator
 {
     public static function generateDocs($project)
     {
-        $path   = "l5-swagger.projects." . $project . '.';
+        $runVersion = config("app.run_version");
+        $runArr     = explode(',', $runVersion);
+        if (count($runArr) < 1 || !in_array($project, $runArr)) {
+            throw new \Exception("no run project");
+        }
+        $docsJson      = "{$project}-api-docs.json";
 
-        $appDir = config('l5-swagger.paths.annotations');
-        $docDir = config('l5-swagger.paths.docs');
-
-        $docFile = $docDir . '/' . config($path . 'routes.docs_json');
+        $appDir  = config('l5-swagger.paths.annotations');
+        $docDir  = config('l5-swagger.paths.docs');
+        $docFile = $docDir . '/' . $docsJson;
 
         if (File::exists($docDir)) {
-
             // delete all existing documentation
             if (File::exists($docFile)) {
                 File::delete($docFile);
@@ -28,7 +31,7 @@ class Generator
         }
 
         if (is_writable($docDir)) {
-            self::defineConstants(config( 'l5-swagger.constants') ?: []);
+            self::defineConstants(config('l5-swagger.constants') ?: []);
 
 
             $excludeDirs = config('l5-swagger.paths.excludes');
@@ -36,7 +39,7 @@ class Generator
                 , ['exclude' => $excludeDirs]);
 
             if (config('l5-swagger.paths.base') !== null) {
-                $swagger->basePath = config($path . 'paths.base');
+                $swagger->basePath = config('l5-swagger.paths.base');
             }
 
             $swagger->saveAs($docFile);
